backend/requirements.txt
aniso8601==9.0.1
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
zipp==3.10.0
